+++ mobile/src/api/wsGameListeners.js
@@ -6,6 +6,7 @@ import {
     setGameError,
     resetGame,
     updateNextRoundReady,
+    applyMendikotTeamUpdate,
     applyMendikotTrumpRevealed,
 } from "../redux/slices/game";
 import { notify } from "../redux/slices/alert";
@@ -109,7 +110,7 @@ export function registerGameListeners() {
     };
 
     const onMendikotTeamUpdate = (data) => {
-        store.dispatch(updateGameState(data));
+        store.dispatch(applyMendikotTeamUpdate(data));
     };
 
     const onMendikotTrumpRevealed = (data) => {